+++ docs/changelog.rst
@@ -39,6 +39,9 @@ Detailed list of changes
 
 - A new command :command:`edit-in-kitty` to :ref:`edit_file`
 
+- Allow getting the last non-empty command output easily via an action or
+  remote control (:pull:`4973`)
+
 - Fix a bug that caused :opt:`macos_colorspace` to always be ``default`` regardless of its actual value (:iss:`5129`)
 
 - ssh kitten: Fix bash not being executed as a login shell since kitty 0.25.0 (:iss:`5130`)